Improvements throughout code
- Especially in decimal fraction and big float implementation
- Added Quantize overloads
- FastInteger can now use mutable number objects
- Reduced use of "long" datatype
- A few more test cases
- Removed redundant remarks documentation

+++ DataUtilities.cs
@@ -185,11 +185,12 @@ namespace PeterO {
     public static int WriteUtf8(String str, int offset, int length, Stream stream, bool replace) {
       if ((stream) == null) throw new ArgumentNullException("stream");
       if ((str) == null) throw new ArgumentNullException("str");
-      if ((offset) < 0) throw new ArgumentException("offset" + " not greater or equal to " + "0" + " (" + Convert.ToString((long)(long)(offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
-      if ((offset) > str.Length) throw new ArgumentException("offset" + " not less or equal to " + Convert.ToString((long)(long)(str.Length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((long)(long)(offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
-      if ((length) < 0) throw new ArgumentException("length" + " not greater or equal to " + "0" + " (" + Convert.ToString((long)(long)(length),System.Globalization.CultureInfo.InvariantCulture) + ")");
-      if ((length) > str.Length) throw new ArgumentException("length" + " not less or equal to " + Convert.ToString((long)(long)(str.Length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((long)(long)(length),System.Globalization.CultureInfo.InvariantCulture) + ")");
-      if (((str.Length - offset)) < length) throw new ArgumentException("str's length minus " + offset + " not greater or equal to " + Convert.ToString((long)(long)(length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((long)(long)((str.Length - offset)),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if ((offset) < 0) throw new ArgumentException("offset" + " not greater or equal to " + "0" + " (" + Convert.ToString((offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if ((offset) > str.Length) throw new ArgumentException("offset" + " not less or equal to " + Convert.ToString((str.Length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if ((length) < 0) throw new ArgumentException("length" + " not greater or equal to " + "0" + " (" + Convert.ToString((length),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if ((length) > str.Length) throw new ArgumentException("length" + " not less or equal to " + Convert.ToString((str.Length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((length),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if (((str.Length - offset)) < length) throw new ArgumentException("str's length minus " + offset + " not greater or equal to " + Convert.ToString((length),System.Globalization.CultureInfo.InvariantCulture) + " (" + 
+                                                                        Convert.ToString((str.Length - offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
       byte[] bytes;
       int retval = 0;
       bytes = new byte[StreamedStringBufferLength];
@@ -286,11 +287,11 @@ namespace PeterO {
                                         StringBuilder builder,
                                         bool replace) {
       if ((data) == null) throw new ArgumentNullException("data");
-      if ((offset) < 0) throw new ArgumentException("offset" + " not greater or equal to " + "0" + " (" + Convert.ToString((long)(long)(offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
-      if ((offset) > data.Length) throw new ArgumentException("offset" + " not less or equal to " + Convert.ToString((long)(long)(data.Length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((long)(long)(offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
-      if ((bytesCount) < 0) throw new ArgumentException("bytesCount" + " not greater or equal to " + "0" + " (" + Convert.ToString((long)(long)(bytesCount),System.Globalization.CultureInfo.InvariantCulture) + ")");
-      if ((bytesCount) > data.Length) throw new ArgumentException("bytesCount" + " not less or equal to " + Convert.ToString((long)(long)(data.Length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((long)(long)(bytesCount),System.Globalization.CultureInfo.InvariantCulture) + ")");
-      if (((data.Length - offset)) < bytesCount) throw new ArgumentException("data's length minus " + offset + " not greater or equal to " + Convert.ToString((long)(long)(bytesCount),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((long)(long)((data.Length - offset)),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if ((offset) < 0) throw new ArgumentException("offset" + " not greater or equal to " + "0" + " (" + Convert.ToString((offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if ((offset) > data.Length) throw new ArgumentException("offset" + " not less or equal to " + Convert.ToString((data.Length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((offset),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if ((bytesCount) < 0) throw new ArgumentException("bytesCount" + " not greater or equal to " + "0" + " (" + Convert.ToString((bytesCount),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if ((bytesCount) > data.Length) throw new ArgumentException("bytesCount" + " not less or equal to " + Convert.ToString((data.Length),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString((bytesCount),System.Globalization.CultureInfo.InvariantCulture) + ")");
+      if (((data.Length - offset)) < bytesCount) throw new ArgumentException("data's length minus " + offset + " not greater or equal to " + Convert.ToString((bytesCount),System.Globalization.CultureInfo.InvariantCulture) + " (" + Convert.ToString(((data.Length - offset)),System.Globalization.CultureInfo.InvariantCulture) + ")");
       if ((builder) == null) throw new ArgumentNullException("builder");
       int cp = 0;
       int bytesSeen = 0;
